requirements.txt
# browser-use 函式庫本身
browser-use

# 用於讀取 .env 檔案
python-dotenv

# Azure OpenAI 和 OpenAI 相容 API
openai

# API 服務相關
fastapi
uvicorn



# 代理和網路相關
requests
aiohttp

# 圖像處理（用於驗證碼識別）
# Pillow
# opencv-python

# 隨機 User-Agent 生成
fake-useragent 